requirements.txt
alabaster==0.7.11
Babel==2.6.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
Django==2.0.7
django-concurrency==1.4
docutils==0.14
idna==2.7
imagesize==1.0.0
Jinja2==2.10
MarkupSafe==1.0
packaging==17.1
Pygments==2.2.0
pyparsing==2.2.0
pytz==2018.5
requests==2.19.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinxcontrib-websupport==1.1.0
urllib3==1.23
